Fixed bug where visualize link was broken
This was because the strong route directive did not pass the routeParams input into the query parameters. The visualize link requires the project/site/region/point ids which are usually stored in the route parameters, by not passing them through, the link was incompatible with the menus.
