+++ src/app/components/visualize/pages/details/details.component.spec.ts
@@ -1,4 +1,4 @@
-import { visualizeRoute } from "@components/visualize/visualize.menus";
+import { visualizeRoute } from "@components/visualize/visualize.routes";
 import { VisualizeModule } from "@components/visualize/visualize.module";
 import { validateBawClientPage } from "@test/helpers/baw-client";
 import { VisualizeComponent } from "./details.component";
